Add files via upload
This Excel file contains a sample of the Sales data which is a flat data set. The Main Objective of this project is to create a Dimensional Model in PowerBI in order to reduce Data redundancy and build the relationship among the data table.  In order to do that first i have to understand the structure of this dataset.

Structure - The Flat data set contains 19 columns and 675368 records. Each record/raw describes a sale order. The columns represent the attributes of a sale order. Each record can be identified as a Fact/Event. In a dimensional data model, the purpose of the Fact Table is to store the measurements/facts. It contains numbers and foreign keys.  The purpose of the dimension table is to store descriptive info about certain areas. So here i can identify the sales order data as a Fact table. Also it has some dimensions such as Customer Dimension, Date Dimension, Geography Dimension and Product Dimension.